SHAP database
To generate the database upon which the ML models are trained, samples of PG2 and PG3 are drawn independently from a uniform distribution in the range [0, 500] MW for a total of 1001 generation scenarios. For the sake of this case study, but without loss of generality, the IEEE 9-bus network is used and demand is assumed constant at 315 MW. PG1 acts as the slack bus.